Metrics mutex should be configurable
While the metrics mutex lock behavior protects any system scraping
metrics from scraping at the same time as the metrics are updated, it
also prevents any scrapes for the whole period that the metrics are
updated. For openstack instances where the API is rather slow, this can
take a very long time.

This enables users to opt out of the metrics mutex safety net, but with
the benefit that the read path on /metrics is always available.