load balancer data should be fetched before reset
In order to update the load balancer metrics the data should be fetched
before resetting the metrics. If there are many load balancers or if the
openstack API is slow to respond the data gathering can take quite a
long time. This allows the metrics to still have the old values until
the new ones are calculated before resetting them.